refactor: ファイル読み込みを簡素化
